Empty groups that have been removed from the configuration
Previously, groups that have been removed from the configuration retain all
their prior members, including users that still exist and can observe their
membership. This seems undesirable and inconsistent with the behaviour of
removed user accounts, which are locked instead. Instead such groups are
emptied now. Special care needs to be taken to not accidentally apply this
new behaviour to the implicitly created primary group of a user with the same
name in case no explicit primary group was specified.