feat: introduce optional stateful user management
In non-nixos settings, it is frequent that some users/groups are created
by default by the distribution, or impurely by the user. Because
userborn is stateless, those users/groups are disabled when they are not
included in the userborn configuration. This change introduce a stateful
mode, activated by the USERBORN_STATEFUL environement variable, where
userborn maintain a list of users/groups it owns and only act on
entities of that list.